flastsm/requirements.txt
# Requirements for flastsm
slack-sdk>=3.20.0
python-dotenv>=0.21.0
requests>=2.28.0
